ai_research_generator/requirements.txt
requests==2.31.0
Jinja2==3.1.2
Markdown==3.5.1
python-dotenv==1.0.0 